+++ pkg/git/sparse_checkout.go
