+++ pkg/settings/template_value_test.go
