+++ pkg/settings/test-fixtures/services_test.yaml
@@ -1,0 +1,14 @@
+settings:
+  kubeconfig: test-fixtures/kubeconfig.yml
+  output: target/test-output
+  sleep: 1s
+  retry-sleep: 250ms
+  retry-count: 3
+  template-values:
+    clusterDomain: unit-test.example.org
+services:
+  - repo: repos/bish
+  - repo: repos/bash
+    branch: special
+  - repo: repos/bosh
+    path: //deployment/foo
