+++ pkg/templates/parse_manifests.go
